requirements.txt
speedtest-cli==2.1.1
python-dotenv==0.10.3